\name{CIhdq-package}
\alias{CIhdq-package}
\alias{CIhdq}
\docType{package}
\title{
\packageTitle{CIhdq}
}
\description{
\packageDescription{CIhdq}
}
\details{
\tabular{ll}{
Package: \tab CIhdq\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2019-03-13\cr
License: \tab GPL-2\cr
}
}
\references{
Feng, X., Huang, J. and Liu, X. (2019). Regularized projection score estimation of treatment effects in high-dimensional quantile regression. Manuscript.
}

\keyword{ package }

