\name{CIhdq-package}
\alias{CIhdq-package}
\alias{CIhdq}
\docType{package}
\title{
Regularized projection score estimation of treatment effects in high-dimensional quantile regression
}
\description{
A regularized projection score method is proposed for estimating treatment effects in quantile regression 
in the presence of high-dimensional confounding covariates. This method is based on an estimated projection 
score function of the low-dimensional treatment parameters in the presence of high-dimensional confounding covariates. 
We propose one-step algorithm and a reffitted wild bootstrapping approach for variance estimation. 
This enables us to construct confidence intervals for the treatment effects in the high-dimensional circumstances.
}
\details{
\tabular{ll}{
Package: \tab CIhdq\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2019-12-12\cr
License: \tab GPL-2\cr
}
}
\references{
Feng, X., Huang, J. and Liu, X. (2019). Regularized projection score estimation of treatment effects in high-dimensional quantile regression. Manuscript.
}

\keyword{ package }

